banner8 수정중

+++ customer_app/views.py
@@ -48,11 +48,16 @@ def ask_question(request):
         form = QuestionForm()
     return render(request, 'create_qna.html', {'form': form})
 
+from django.shortcuts import get_object_or_404
+
 def qna_detail(request, id):
     question = get_object_or_404(Question, pk=id)
 
-    if question.is_answered:
-        answer = get_object_or_404(Answer, question=question)
+    # 답변이 여러 개인 경우, 적절한 조건으로 필터링하여 가져옵니다.
+    answers = Answer.objects.filter(question=question)
+
+    if question.is_answered and answers.exists():
+        answer = answers.first()  # 필요에 따라 조건에 맞게 선택
     else:
         answer = "답변 대기중"
 
@@ -101,14 +106,27 @@ def answer_update(request, id):
     return render(request, 'answer_update.html', {'form': form, 'answer': answer})
 
 def answer_delete(request, id):
-    answer = get_object_or_404(Answer, question=id)
+    # 답변을 가져올 때, question=id로 가져오는데, 답변은 하나의 질문에만 속합니다.
+    # 그러므로 get_object_or_404(Answer, question=id) 대신 get_object_or_404(Answer, id=id)을 사용해야 합니다.
+    answer = get_object_or_404(Answer, id=id)
 
-    if request.user.is_superuser:
-        answer.delete()
-        # 해당 질문의 is_answered 값을 0으로 업데이트
-        question = answer.question
-        question.is_answered = 0
-        question.save()
-        return redirect('qna_detail', id=answer.question.id)
+    if request.user.is_authenticated:
+        # 관리자는 언제나 삭제 권한을 가집니다.
+        if request.user.is_superuser:
+            answer.delete()
+            # 해당 질문의 is_answered 값을 0으로 업데이트
+            question = answer.question
+            question.is_answered = 0
+            question.save()
+            return redirect('qna_detail', id=question.id)
+
+        # 작성자인 경우에만 삭제 권한을 가집니다.
+        elif request.user == answer.user:
+            answer.delete()
+            # 해당 질문의 is_answered 값을 0으로 업데이트
+            question = answer.question
+            question.is_answered = 0
+            question.save()
+            return redirect('qna_detail', id=question.id)
 
     return render(request, 'error.html', {'error_message': '삭제 권한이 없습니다.'})
